+++ front/lib/api/tables.ts
@@ -10,6 +10,7 @@ import { CoreAPI, Err, isSlugified, Ok } from "@dust-tt/types";
 import { parse } from "csv-parse";
 import { DateTime } from "luxon";
 
+import { guessDelimiter } from "@app/lib/api/csv";
 import { AgentTablesQueryConfigurationTable } from "@app/lib/models/assistant/actions/tables_query";
 import logger from "@app/logger/logger";
 
@@ -230,39 +231,7 @@ export async function upsertTableFromCsv({
 async function rowsFromCsv(
   csv: string
 ): Promise<Result<CoreAPIRow[], CsvParsingError>> {
-  // Detect the delimiter: try to parse the first 2 lines with different delimiters,
-  // keep the one that works for both lines and has the most columns.
-  let delimiter: string | undefined = undefined;
-  let delimiterColsCount = 0;
-  for (const d of [",", ";", "\t"]) {
-    const records: unknown[][] = [];
-    try {
-      const parser = parse(csv, { delimiter: d });
-      for await (const record of parser) {
-        records.push(record);
-        if (records.length == 2) {
-          break;
-        }
-      }
-    } catch (e) {
-      // Ignore error.
-      continue;
-    }
-
-    const [firstRecord, secondRecord] = records;
-    // Check for more than one line to ensure sufficient data for accurate delimiter detection.
-    if (!secondRecord) {
-      continue;
-    }
-
-    if (!!firstRecord.length && firstRecord.length === secondRecord.length) {
-      if (firstRecord.length > delimiterColsCount) {
-        delimiterColsCount = firstRecord.length;
-        delimiter = d;
-      }
-    }
-  }
-
+  const delimiter = await guessDelimiter(csv);
   if (!delimiter) {
     return new Err({
       type: "invalid_delimiter",
